+++ x/vpool/keeper/pool_state.go
@@ -22,16 +22,16 @@ func (k Keeper) CreatePool(
 	maintenanceMarginRatio sdk.Dec,
 	maxLeverage sdk.Dec,
 ) {
-	k.savePool(ctx, types.NewVPool(
-		pair,
-		tradeLimitRatio,
-		quoteAssetReserve,
-		baseAssetReserve,
-		fluctuationLimitRatio,
-		maxOracleSpreadRatio,
-		maintenanceMarginRatio,
-		maxLeverage,
-	))
+	k.savePool(ctx, &types.VPool{
+		Pair:                   pair,
+		BaseAssetReserve:       baseAssetReserve,
+		QuoteAssetReserve:      quoteAssetReserve,
+		TradeLimitRatio:        tradeLimitRatio,
+		FluctuationLimitRatio:  fluctuationLimitRatio,
+		MaxOracleSpreadRatio:   maxOracleSpreadRatio,
+		MaintenanceMarginRatio: maintenanceMarginRatio,
+		MaxLeverage:            maxLeverage,
+	})
 	k.SaveSnapshot(ctx, pair, quoteAssetReserve, baseAssetReserve)
 }